refactor(knowledge): 将 Dashboard 二级菜单全量重命名为 Pipelines
对齐之前 PR 未处理干净的命名遗留，涵盖导航标签、路由路径、API endpoint、
Schema 类型名、前端组件/变量名、测试断言与 E2E 截图的系统性替换。

后端：/dashboard → /pipelines/overview，DashboardResponse → PipelinesResponse
前端：KnowledgeNav 标签、page.tsx 迁移、API client 函数/类型重命名
测试：KnowledgeNav / PipelinesPage 单元测试、helpers mock、E2E smoke
文档：user-guide.md、knowledges.md 中 Knowledge Dashboard 引用
