--- apps/negentropy-ui/app/api/knowledge/dashboard/route.ts
@@ -1,5 +1,0 @@
-import { proxyGet } from "../_proxy";
-
-export async function GET(request: Request) {
-  return proxyGet(request, "/knowledge/dashboard");
-}
