+++ apps/negentropy-ui/app/api/knowledge/pipelines/overview/route.ts
@@ -1,0 +1,5 @@
+import { proxyGet } from "../../_proxy";
+
+export async function GET(request: Request) {
+  return proxyGet(request, "/knowledge/pipelines/overview");
+}
