+++ apps/negentropy-ui/tests/e2e/smoke.spec.ts-snapshots/knowledge-pipelines-runs-dark-chromium.png
