+++ apps/negentropy-ui/tests/unit/ui/KnowledgeNav.test.tsx
@@ -4,7 +4,7 @@ import { describe, expect, it, vi } from "vitest";
 import { KnowledgeNav } from "@/components/ui/KnowledgeNav";
 
 vi.mock("next/navigation", () => ({
-  usePathname: () => "/knowledge/dashboard",
+  usePathname: () => "/knowledge/pipelines",
 }));
 
 vi.mock("@/components/providers/NavigationProvider", () => ({
@@ -14,15 +14,15 @@ vi.mock("@/components/providers/NavigationProvider", () => ({
 }));
 
 describe("KnowledgeNav", () => {
-  it("保留 Dashboard 二级导航项并高亮 dashboard 页面", () => {
-    render(<KnowledgeNav title="Dashboard" />);
+  it("保留 Pipelines 二级导航项并高亮 pipelines 页面", () => {
+    render(<KnowledgeNav title="Pipelines" />);
 
-    const dashboardLink = screen.getByRole("link", { name: "Dashboard" });
+    const pipelinesLink = screen.getByRole("link", { name: "Pipelines" });
     const baseLink = screen.getByRole("link", { name: "Knowledge Base" });
 
-    expect(dashboardLink).toBeInTheDocument();
-    expect(dashboardLink).toHaveAttribute("href", "/knowledge/dashboard");
-    expect(dashboardLink.className).toContain("bg-foreground");
+    expect(pipelinesLink).toBeInTheDocument();
+    expect(pipelinesLink).toHaveAttribute("href", "/knowledge/pipelines");
+    expect(pipelinesLink.className).toContain("bg-foreground");
     expect(baseLink).toHaveAttribute("href", "/knowledge/base");
   });
 });
