+++ apps/negentropy/src/negentropy/knowledge/schemas.py
@@ -277,7 +277,7 @@ class SearchRequest(BaseModel):
     metadata_filter: dict[str, Any] | None = None
 
 
-class DashboardResponse(BaseModel):
+class PipelinesResponse(BaseModel):
     corpus_count: int
     knowledge_count: int
     last_build_at: str | None = None
