+++ docs/user-guide.md
@@ -513,14 +513,14 @@ PLAYWRIGHT_AUTH_MODE=dev-cookie PLAYWRIGHT_AUTH=1 \
 | 导航项        | 路径                   | 功能                   |
 | :------------ | :--------------------- | :--------------------- |
 | **Base**      | `/knowledge/base`      | 语料库管理（核心入口） |
-| **Dashboard** | `/knowledge/dashboard` | 知识库仪表盘           |
+| **Pipelines** | `/knowledge/pipelines` | 管线运行概览           |
 | **Catalog**   | `/knowledge/catalog`   | 知识目录（树形结构）   |
 | **Documents** | `/knowledge/documents` | 全局文档列表           |
 | **Graph**     | `/knowledge/graph`     | 知识图谱可视化         |
 | **Pipelines** | `/knowledge/pipelines` | Pipeline 监控          |
 | **APIs**      | `/knowledge/apis`      | API 文档与测试         |
 
-### 4.2 Dashboard 仪表盘
+### 4.2 Pipelines 管线概览
 
 仪表盘提供知识库的全局概览，采用两栏布局：
